Added matrix tools to camera, tooltip in editor
Added tools in the camera class to allow you to get the (predicted)
matrix of the camera at any given time.  This is used in the map class
to translate coordinates from world space to screen space, and back.
This is used to have a tooltip in the editor when rolling over entities.